fix: internalize raw cluster frame execution
Keep raw RESP routing out of the public cluster client facade while retaining it in the cluster sublibrary internal module. Add deterministic frame-identity coverage for redirects, transient failures, topology refreshes, and transport reconnects.


Copilot-Session: f243bf49-7fc9-44b4-ae74-a6a9311af2fa
