+++ hask-redis-mux/hask-redis-mux.cabal
@@ -116,6 +116,7 @@ library cluster
     hs-source-dirs:   lib/cluster
     exposed-modules:  Database.Redis.Cluster
                     , Database.Redis.Cluster.Client
+                    , Database.Redis.Cluster.Internal.RawCommand
                     , Database.Redis.Cluster.Commands
                     , Database.Redis.Cluster.SlotMapping
                     , Database.Redis.Cluster.ConnectionPool
@@ -124,6 +125,7 @@ library cluster
                     , Database.Redis.Internal.Multiplexer
                     , Database.Redis.Cluster.Internal.Topology
                     , Database.Redis.Standalone
+    other-modules:    Database.Redis.Cluster.Internal.ClientImplementation
     build-depends:    attoparsec >=0.14 && <0.15
                     , bytestring >=0.12 && <0.13
                     , client
@@ -194,7 +196,8 @@ test-suite PublicApiSpec
     import:           test-defaults
     type:             exitcode-stdio-1.0
     main-is:          PublicApiSpec.hs
-    build-depends:    redis
+    build-depends:    directory >=1.3 && <1.4
+                    , redis
 
 test-suite ClusterSpec
     import:           test-defaults
