+++ hask-redis-mux/lib/cluster/Database/Redis/Cluster/Internal/RawCommand.hs
@@ -1,0 +1,9 @@
+-- | Internal cluster-sublibrary API for protocol adapters that forward
+-- pre-parsed RESP frames without altering their encoded representation.
+module Database.Redis.Cluster.Internal.RawCommand
+  ( RawClusterRoute (..)
+  , executeRawClusterCommand
+  , executeRawClusterCommandUsingDelay
+  ) where
+
+import           Database.Redis.Cluster.Internal.ClientImplementation
