+++ hask-redis-mux/test/PublicApiSpec.hs
@@ -5,7 +5,9 @@
 module Main (main) where
 
 import           Control.Exception (try)
+import           Control.Monad     (filterM)
 import           Database.Redis
+import           System.Directory  (doesFileExist)
 import           Test.Hspec
 
 main :: IO ()
@@ -13,6 +15,22 @@ main = hspec $ describe "Database.Redis timeout-aware public API" $ do
   it "exports the migration-compatible ordinary cluster error" $ do
     RedisCommandError "ERR full server cause"
       `shouldBe` RedisCommandError "ERR full server cause"
+
+  it "keeps raw cluster frame execution out of public facades" $ do
+    clientSource <- findSource
+      [ "lib/cluster/Database/Redis/Cluster/Client.hs"
+      , "hask-redis-mux/lib/cluster/Database/Redis/Cluster/Client.hs"
+      ]
+    cabalSource <- findSource
+      [ "hask-redis-mux.cabal"
+      , "hask-redis-mux/hask-redis-mux.cabal"
+      ]
+    publicExports <- takeWhile (/= ") where") . lines <$> readFile clientSource
+    unlines publicExports `shouldNotContain` "executeRawClusterCommand"
+    unlines publicExports `shouldNotContain` "RawClusterRoute"
+    topLevelReexports <- takeWhile (/= "library resp") . lines <$> readFile cabalSource
+    unlines topLevelReexports `shouldNotContain`
+      "Database.Redis.Cluster.Internal.RawCommand"
 
   it "exports redaction-safe cluster authentication policies" $ do
     let createAuthenticated
@@ -56,3 +74,10 @@ assertImmediateTimeout endpoint = \case
     connectionTimeoutSeconds timeoutError `shouldBe` 0
   Right _ ->
     expectationFailure "timeout-aware helper unexpectedly connected"
+
+findSource :: [FilePath] -> IO FilePath
+findSource candidates = do
+  matches <- filterM doesFileExist candidates
+  case matches of
+    path : _ -> return path
+    [] -> expectationFailure "Could not locate public API source" >> fail "unreachable"
